Split bencode parser in modules
